feat: basic navbar styling
